requirements.txt
langchain
langchain-google-genai
langchain-chroma
google-genai
python-dotenv
chromadb